gpl-asm v0.8.1: fix text-parser byte-length accounting
The 'known-flaky debug_assert_eq in lib.rs:91' was a real bug, not
flakiness: parse.rs rebuilt DisasmResult.total_bytes from per-
instruction length estimates that disagreed with the encoder in
three places (none affected encoded bytes, which stayed corpus-
byte-identical throughout):

- RetVal + nested gpl_search: estimate dropped inner_raw_tail and
  counted params past [0] the encoder never writes.
- top-level gpl_search: params past [0] are byte-for-byte inside
  raw_tail; counting both double-counted them.
- gpl_setrecord: the encoder writes the access_complex body with
  no leading dispatch byte, but expression_byte_len charges one
  per ComplexAccess token.

text_roundtrip now asserts parsed.total_bytes == src.len() per
chunk so estimate drift fails in release builds too. cargo test
passes in debug; CLAUDE.md's --release workaround note removed.
Also: collapse a nested if in bin/gpl-asm.rs and de-nest two
format! calls in corpus_roundtrip.rs (clippy 1.96).

+++ tools/gpl-asm/src/lib.rs
@@ -337,7 +337,7 @@ pub fn encode_expression(out: &mut Vec<u8>, instr_offset: usize, expr: &Expressi
             // and lo = ((bytes[2] << 8) | bytes[3]) as u16.
             // Inverse: split value into hi (top 16 bits, signed) and
             // lo (bottom 16 bits, unsigned).
-            let hi = ((*value as i32 >> 16) as u16).to_be_bytes();
+            let hi = ((*value >> 16) as u16).to_be_bytes();
             let lo = (*value as u32 & 0xFFFF) as u16;
             let lo = lo.to_be_bytes();
             out.extend_from_slice(&hi);
